[Spring 장바구니 - 2단계] 조동현(후디) 미션 제출합니다. (#130)
* docs: 2단계 기능 구현 목록 작성

* feat: 상품 도메인에 설명, 재고수량 필드 추가

프론트엔드 크루와의 회의 결과 반영

* refactor: 상품을 찾을 수 없을 경우에 대한 예외 추가

- 상품을 찾을 수 없는 경우 400 대신 404를 응답하도록 변경
- 인수 테스트 코드 추가

* docs: 기능 구현 목록 반영

* feat: 액세스 토큰에서 유저 아이디를 추출하는 Argument Resolver 구현

* refactor: Cart Item API가 유저를 식별할 때 customerName 대신 액세스 토큰으로 식별하도록 변경

* refactor: Order API가 유저를 식별할 때 customerName 대신 액세스 토큰으로 식별하도록 변경

* docs: 2단계 기능 구현 목록 추가

* test: 장바구니 및 주문 API에 인가 테스트 추가 (토큰 만료, 토큰 유효하지 않음)

* docs: 2단계 기능 구현 목록 갱신

테스트코드 Display Name 오타 수정

* feat: 변경된 명세에 따라 카트 아이템 추가 및 조회 기능 수정

* feat: 특정 상품이 특정 유저의 장바구니에 존재하는지 여부를 조회하는 API 구현

* refactor: dto를 요청과 응답에 따라 패키지 분리

* feat: 장바구니 수정 API 기능 구현

* docs: 2단계 기능 구현 목록 변경

* feat: 장바구니 추가 시 존재하지 않은 제품 ID를 전달받은 경우 404에러가 발생하도록 구현

* feat: 장바구니 항목 삭제 시 존재하지 않은 장바구니 항목 ID가 전달되면 404 에러가 발생하도록 수정

* docs: 기능 개선 목록 갱신

* feat: 주문 추가, 조회 기능을 변경된 기획에 맞게 수정

* feat: 카트 아이템 수정 시 요청한 카트 아이템의 productId 와, Body 의 productId 가 다를 경우 400 에러가 발생하도록 구현

* refactor: Dao에 대한 인터페이스 추가

* refactor: ResponseEntity에 제네릭 타입 추가

* refactor: cartId를 cartItemId로 변경

* refactor: single_order 테이블을 orders로, order_product 테이블을 order_detail로 변경

* refactor: SQL내의 대문자로 되어있는 테이블이름을 소문자로 변경

* refactor: schema.sql 에서 외래키를 설정할 때 alter table 대신 create table 에서 외래키를 설정하도록 변경

* chore: mysql-connector-java 의존성 추가

* fix: 테이블명을 소문자로 변경

* refactor: 불필요한 출력 로직 제거

* feat: 회원 요청과 응답 DTO의 depth 제거

* feat: zoneCode와 ZoneCode를 zonecode, Zonecode로 변경

* feat: 전체 장바구니 조회시 ID 필드를 id에서 cartItemId로 변경

* feat: 이미 장바구니에 담긴 상품을 또다시 등록하면 400에러가 발생하도록 구현

* feat: 에러 반환을 문자열 대신 ErrorResponse로 하도록 개선

* feat: 회원가입 시 이미 이메일이 존재하는 경우 400 에러 발생

* refactor: API 문서에 따라 토큰 발급 응답값의 회원 ID필드를 customerId에서 userId로 변경

* fix: preflight 대응 위해 AuthInterceptor preHandle 메소드 수정

* refactor: JdbcCartItemDao에 SimpleJdbcInsert 사용하여 코드 개선

* refactor: CustomerDao의 조회 메서드에서 EmptyResultDataAccessException을 CustomerNotFoundException으로 예외를 번역

* test: 인수 테스트 최적화

인수 테스트에서 DirtiesContext 대신 Sql 어노테이션을 사용하여 테스트 격리

* refactor: OrdersDao와 OrderDetailDao 분리

OrderDao를 OrdersDao로 이름 변경

* refactor: JdbcOrdersDao 리팩토링

메소드 추출하여 가독성 개선
OrdersNotFoundException 추가

* refactor: JdbcProductDao 리팩토링

RowMapper를 상수로 분리
SimpleJdbcInsert 사용하여 코드 개선

* refactor: auth 패키지와 shoppingcart 패키지의 Config를 병합

* docs: 기능 구현 목록 개선

* feat: schema에 on delete 제약 조건 추가

* fix: 빈 BirthDay의 value를 null로 설정

* refactor: ProductRepository 구현

* refactor: OrderRepository 구현

* refactor: CustomerRepository 구현

* refactor: CartItemRepository 구현

* refactor: 단건 조회시 DAO가 Optional을 반환하도록 개선

* refactor: 일관적이지 않게 사용된 final 키워드 제거 (파라미터 및 지역변수)

* refactor: final 키워드 추가

* refactor: 사용되지 않는 코드 제거

* refactor: Id 도메인 제거

Customer Id 타입을 Integer에서 Long으로 변경

* refactor: Spring Security 의존성 제거

Sha256 암호화를 위한 PasswordEncoder 구현

* refactor: RowMapper를 상수로 변경

* refactor: 컬럼이름을 상수로 분리

* refactor: DTO의 생성자 파라미터로 도메인 객체를 전달하여 DTO를 생성할 수 있도록 개선

+++ src/main/java/woowacourse/shoppingcart/controller/CustomerController.java
@@ -11,9 +11,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import woowacourse.shoppingcart.dto.CustomerRequest;
-import woowacourse.shoppingcart.dto.CustomerResponse;
-import woowacourse.shoppingcart.dto.EmailDuplicationResponse;
+import woowacourse.shoppingcart.dto.request.CustomerRequest;
+import woowacourse.shoppingcart.dto.response.CustomerResponse;
+import woowacourse.shoppingcart.dto.response.EmailDuplicationResponse;
 import woowacourse.shoppingcart.service.CustomerService;
 
 @RestController
@@ -27,7 +27,7 @@ public class CustomerController {
 
     @PostMapping("/customers")
     public ResponseEntity<Void> createCustomer(@RequestBody CustomerRequest customerRequest) {
-        int id = customerService.create(customerRequest);
+        long id = customerService.create(customerRequest);
         return ResponseEntity.created(URI.create("/api/customers/" + id)).build();
     }
